+++ wemake_python_styleguide/visitors/complexity/counts.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 
-# TODO: count the number of functions per file/class, classes per file
-
 import ast
+from collections import defaultdict
+from typing import DefaultDict
 
 from wemake_python_styleguide.errors.complexity import (
+    TooManyMethodsViolation,
     TooManyModuleMembersViolation,
 )
 from wemake_python_styleguide.logics.functions import is_method
@@ -13,35 +14,26 @@ from wemake_python_styleguide.types import ModuleMembers
 from wemake_python_styleguide.visitors.base.visitor import BaseNodeVisitor
 
 
-class _MembersCounter(object):
-    """Helper class to encapsulate logics from the visitor."""
-
-    def __init__(self, delegate: BaseNodeVisitor) -> None:
-        self.delegate = delegate
-        self.public_items_count = 0
-
-    def check_members_count(self, node: ModuleMembers):
-        """This method increases the number of module members."""
-        parent = getattr(node, 'parent', None)
-        is_class_method = is_method(getattr(node, 'function_type', None))
-
-        if isinstance(parent, ast.Module) and not is_class_method:
-            self.public_items_count += 1
-            max_members = self.delegate.options.max_module_members
-            if has_just_exceeded_limit(self.public_items_count, max_members):
-                error = TooManyModuleMembersViolation(
-                    node, text=self.delegate.filename,
-                )
-                self.delegate.add_error(error)
-
-
 class ModuleMembersVisitor(BaseNodeVisitor):
     """Counts classes and functions in a module."""
 
     def __init__(self, *args, **kwargs) -> None:
         """Creates a counter for tracked metrics."""
         super().__init__(*args, **kwargs)
-        self._counter = _MembersCounter(self)
+        self._public_items_count = 0
+
+    def _check_members_count(self, node: ModuleMembers):
+        """This method increases the number of module members."""
+        parent = getattr(node, 'parent', None)
+        is_real_method = is_method(getattr(node, 'function_type', None))
+
+        if isinstance(parent, ast.Module) and not is_real_method:
+            self._public_items_count += 1
+            max_members = self.options.max_module_members
+            if has_just_exceeded_limit(self._public_items_count, max_members):
+                self.add_error(
+                    TooManyModuleMembersViolation(node, text=self.filename),
+                )
 
     def visit_ClassDef(self, node: ast.ClassDef):
         """
@@ -51,7 +43,7 @@ class ModuleMembersVisitor(BaseNodeVisitor):
             TooManyModuleMembersViolation
 
         """
-        self._counter.check_members_count(node)
+        self._check_members_count(node)
         self.generic_visit(node)
 
     def visit_FunctionDef(self, node: ast.FunctionDef):
@@ -62,5 +54,34 @@ class ModuleMembersVisitor(BaseNodeVisitor):
             TooManyModuleMembersViolation
 
         """
-        self._counter.check_members_count(node)
+        self._check_members_count(node)
+        self.generic_visit(node)
+
+
+class MethodMembersVisitor(BaseNodeVisitor):
+    """Counts methods in a single class."""
+
+    def __init__(self, *args, **kwargs) -> None:
+        """Creates a counter for tracked methods in different classes."""
+        super().__init__(*args, **kwargs)
+        self._methods: DefaultDict[ast.ClassDef, int] = defaultdict(int)
+
+    def _check_method(self, node: ast.FunctionDef):
+        parent = getattr(node, 'parent', None)
+        is_real_method = is_method(getattr(node, 'function_type', None))
+        if isinstance(parent, ast.ClassDef) and is_real_method:
+            self._methods[parent] += 1
+            max_methods = self.options.max_methods
+            if has_just_exceeded_limit(self._methods[parent], max_methods):
+                self.add_error(TooManyMethodsViolation(node, text=parent.name))
+
+    def visit_FunctionDef(self, node: ast.FunctionDef):
+        """
+        Counts the number of methods in a single class.
+
+        Raises:
+            TooManyMethodsViolation
+
+        """
+        self._check_method(node)
         self.generic_visit(node)